+++ src/components/list_box_table/render_group_label.jsx
@@ -1,14 +1,14 @@
-import { classPrefix } from '../../constants/class_prefix';
+import { makeBEMClass } from '../../helpers/make_bem_class';
 
 export function renderGroupLabel(props, componentState, componentProps) {
-  const { columns, renderTableGroupRow, renderTableGroupHeaderCell } = componentProps;
+  const { classPrefix, columns, renderTableGroupRow, renderTableGroupHeaderCell } = componentProps;
 
   return renderTableGroupRow({
-    className: `${classPrefix}listbox__table-group-row`,
+    className: makeBEMClass(classPrefix, 'table-group-row'),
     children: renderTableGroupHeaderCell({
       ...props,
       colSpan: Object.keys(columns).length,
-      className: `${classPrefix}listbox__table-group-header`,
+      className: makeBEMClass(classPrefix, 'table-group-header'),
     }, componentState, componentProps),
   }, componentState, componentProps);
 }