+++ src/helpers/make_bem_class.js
@@ -1,0 +1,7 @@
+export function makeBEMClass(prefix, ...values) {
+  if (!prefix) {
+    return null;
+  }
+
+  return values.map((value) => `${prefix}__${value}`).join(' ') || prefix;
+}